swap lambdas so accidental trigger doesn't try to do anything useful
